LOK: don't create the dialog when loading new views
By default dialogs/decks are re-created when a new view is
created/attached, if they are visible. For the modified dialog
this is unexpected, as the user doesn't expect to be doing
actions upon loading a document, just because another
user was at the time had those dialog open. Currently the suppression
is for LOK only.

This also adds support to suppress the recreation of any
dialog by flagging it with the SfxChildWindowFlags::NEVERCLONE
flag upon invoking RegisterChildWindow.

similar to c9829f585103f3daa9749c5b0f479297edd42042

Change-Id: I50598dbbc7a0f156fc0b0d0777ef0fb350610281
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/176923
Reviewed-by: Caolán McNamara <caolan.mcnamara@collabora.com>
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>