cool#10610 Ensure the parent-child relations of comments.
Previously, parent-child (comment-reply) relations were established using the position of comments.
For some time now, users can add comments to the same position and they still won't be replies to first comment.
This is provided with parent-name, parentPostItId etc variables.
When copying a portion of text, comments are also copied.
While copying the comments, we need to re-establish their parent-child relations.
This code uses a temp map for the purpose and re-connects replies to their new parent comments.

Signed-off-by: Gökay Şatır <gokaysatir@collabora.com>
Change-Id: Id0fe7f419228b115a6e3d1e10ccb2fc70db57cea
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/177396
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>