pdf: Rearrange general tab of PDF dialog, add PDF vers. selection
Repurpuses PDF/A combobox to also add PDF 1.7 and PDF 2.0 as
options, as those are exclusive (either you have classic PDF, or
you select PDF/A which is a special version of some PDF version.
No need to have a check box for PDF/A anymore, so this was removed.

- Moved the version selection to top
- Added "Forms" frame, which for all the form options.
- Moved Tagged PDF into "Structure" frame.
- Reworded some text (extended tips, tooltips) for PDF versions

Change-Id: I731222ee02afea6a49e4213d18f27d1beae9da10
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/178270
Tested-by: Miklos Vajna <vmiklos@collabora.com>
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>