enable spotlight subset in kit mode
Use the index of the style within the list of writer's instantiated
styles as the indicator number to use, so this is stable regardless of
what particular mode the style dialog/panel is using.

Unused styles don't get a number indicator, but do have a (stable) color
until they are used in the document.

For kit, Para/Char style mode renders all styles, not just the subset
selected in the style dialog/panel, when Para/Char spotlight is enabled.

Change-Id: I513e4a676a7c4f7f542c016afa5a79d158778210
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/178481
Reviewed-by: Caolán McNamara <caolan.mcnamara@collabora.com>
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Michael Meeks <michael.meeks@collabora.com>