undo/redo of comment insertion assumes Par2 content matches TextObject
contents (as much as conversion to plain text makes possible)

To make things easier, add an EditTextObject::GetText() like
EditEngine::GetText() that serializes multiple paragraphs in an
equivalent way so there isn't a need to create an editengine just
to do this conversion.

Modify and extend DesktopLOKTest::testCommentsCallbacksWriter to
use a html payload and add a undo/redo to it.

Change-Id: I4d895138e919bab54ebbbcb966f9b9faef574086
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/179594
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>