cool#10630 doc sign: fix Impress sign line, to be able to finish signing again
Open a PDF for signing on the desktop, insert a signature line, click
the "finish signing" button on the infobar, Draw first asks if you want
to save your modified document as ODG, then crashes.

The first problem is that opening a PDF for signing and adding a
signature line / visual signature didn't result in a modified doc model,
but that now happens since commit
aca32a55456aa4e907b216fb490b3c15d26c3d55 (tdf#146547 sfx2: allow
read-only documents to be modified, 2023-06-16). Probably we want to
keep this feaature that read-only documents can be modified in some
cases, so explicitly mark the PDF document as not-modified before
signing in SfxObjectShell::SignDocumentContentUsingCertificate().

The other problem is that visual signing tried to reload the document:
this is handy, as it shows if our PDF export worked correctly, but
reloading is problematic since the signing code moved to async and it
assumes that the underlying object shell is never deleted, so callbacks
can have a pointer to it without worrying about the lifecycle. Fix this
problem by not reloading, but clearing the marker on the signature shape
and invalidating the signature status, instead of reloading, in
SfxObjectShell::ExecFile_Impl().

With this, visual signing of an existing PDF file works again in desktop
Draw.

Change-Id: I743983947d4607aa06674a1329bc8efb5af8a6d9
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/179821
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Caolán McNamara <caolan.mcnamara@collabora.com>