sd pdfium: add page number when saving/loading pdf pages as images
When a pdf file is opened with pdfium and then saved as odg, each page
is saved as a draw:image of the pdf. But since no page number info was
given, when opening the odg file, all the pages are shown as the first
one.

This adds a loext:page-number to the saved odg and the logic to load a
specific page from an embedded pdf.

Added loadGraphicAtPage to XGraphicStorageHandler.

Change-Id: I398b7ad7f03cc7fb930a4e7b7620c6f238b41c0d
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/180815
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>