Setting zoom must also reset offsets in LOK cached view
The problem was, that having a spreadsheet with objects,
e.g. rectangles, drawn e.g. over M15:O31, and zooming
the spreadsheet in Online, the object was painted with
substantial offset from the actual placement (it could
be three rows below, and some millimeters to the right).

The problem was, that respective ViewObjectContact did
not reset its maGridOffset upon changing zoom, because
it belonged to a view cached in ScGridWindow.

It is different from what GetSdrPageView will return in
ScDrawView::resetGridOffsetsForAllSdrPageViews; we need
to handle that in ScModelObj::setClientZoom, or zooming
in Calc would paint drawing objects in wrong positions.

The sizes of the objects are still painted incorrectly.

Change-Id: Ie2f90568f0570bc0492b614a5fa9128c3e5e9a1b
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/181611
Tested-by: Jenkins
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>
(cherry picked from commit a383d7afa4e82bea728ecddf1bc44bbd9fd0eddb)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/181618
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>